v4l2-subdev: add VIDIOC_SUBDEV_QUERYCAP ioctl
While normal video/radio/vbi/swradio nodes have a proper QUERYCAP ioctl
that apps can call to determine that it is indeed a V4L2 device, there
is currently no equivalent for v4l-subdev nodes. Adding this ioctl will
solve that, and it will allow utilities like v4l2-compliance to be used
with these devices as well.

SUBDEV_QUERYCAP currently returns the version, device_caps and the
entity_id of the subdevice (if there is any).

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>
[Kieran: Fix entity.parent and entity.id -> graph_obj]
Signed-off-by: Kieran Bingham <kieran.bingham+renesas@ideasonboard.com>