media: i2c: adv748x: add adv748x driver
Provide basic support for the ADV7481 and ADV7482.

The driver is modelled with 2 subdevices to allow simultaneous streaming
from the AFE (Analog front end) and HDMI inputs.

Presently the HDMI is hardcoded to link to the TXA CSI bus, whilst the
AFE is linked to the TXB CSI bus.

The driver is based on a prototype by Koji Matsuoka in the Renesas BSP,
and an earlier rework by Niklas Söderlund.

Signed-off-by: Kieran Bingham <kieran.bingham+renesas@ideasonboard.com>

---

(Kierans Change log ... This should be deleted)

[LOCAL] media: i2c: adv7482: add CP and SDP controls
[LOCAL] media: i2c: adv7482: add V4L2_CID_PIXEL_RATE control
adv7482: Move to subfolder
adv7482: Refactor into separate .o files.

The hack_is_hdmi() trick is still involved, and two instantiations of
the driver are loaded to create a subdev for each of CVBS and HDMI.

However, this moves us forward to separating the two 'devices' into
their own structures, and providing a better framework to clean up the
subdev usage.

A few bugs are fixed in this squash as well :-)

adv7482: Remove secondary cvbs and hdmi state strutures
adv7482: Report errors on probe
adv7482: Create common subdev helper

Move the subdevs away from being i2c subdevices.

Instead, register the client with the adv7482 state, and register
the subdevs with the device state.

adv7482: Reduce entity pad counts

Pull the entities down to a single source/sink per entity

Signed-off-by: Kieran Bingham <kieran.bingham+renesas@ideasonboard.com>

adv7482: Remove adv7482_media_has_route

'has route' is no longer applicable, as we do not use a single
entity to map all routes through the ADV.

Signed-off-by: Kieran Bingham <kieran.bingham+renesas@ideasonboard.com>

adv7482: hdmi: Use hdmi prefix in all function names

Signed-off-by: Kieran Bingham <kieran.bingham+renesas@ideasonboard.com>

adv7482: hdmi: Fix Sink/Source pads

Now that we have split to only 2 pads, we need to make sure
the usages of the pads are correctly mapped to the right pad
numbers

Signed-off-by: Kieran Bingham <kieran.bingham+renesas@ideasonboard.com>

adv7482: cvbs/sdp: Prefix all functions with sdp

Move away from 'CVBS' which is one of the functions,
to the name of the block. SDP - Presumably Standard Def Processor

Signed-off-by: Kieran Bingham <kieran.bingham+renesas@ideasonboard.com>

adv7482: cvbs/sdp: Fix Sink/Source pads

Now that we have split to only 2 pads, we need to make sure
the usages of the pads are correctly mapped to the right pad
numbers

Signed-off-by: Kieran Bingham <kieran.bingham+renesas@ideasonboard.com>

adv7482: Remove incorrect pad references

These described the state of pads when they were a single
entity. Now that we have split, it's just wrong.

Remove and clean up

adv7482: State authors in -core