v4l: vsp1: Support chained display lists
When display lists are linked in a chain, they will be processed automatically
by the hardware, with each list linking to the next. Only on the last display
list will the frame end interrupt be fired to mark the completion event.

Upon frame-end, the chain will be iterated to release each display list
back to the free list.

Signed-off-by: Kieran Bingham <kieran@bingham.xyz>