v4l: vsp1: Provide a writeback video device
When the VSP1 is used in an active display pipeline, the output of the
WPF can supply the LIF entity directly and simultaneously write to
memory.

Support this functionality in the VSP1 driver, by extending the WPF
source pads, and establishing a V4L2 video device node connected to the
new source.

The source will be able to perform pixel format conversion, but not
rescaling, and as such the output from the memory node will always be
of the same dimensions as the display output.

Signed-off-by: Kieran Bingham <kieran.bingham+renesas@ideasonboard.com>

---
Changes since RFC
 - Fix checkpatch.pl warnings
 - Adapt to use a single source pad for the Writeback and LIF
 - Use WPF properties to determine when to create links instead of VSP
 - Remove incorrect vsp1_video_verify_format() changes
 - Spelling and grammar fixes

v3:
 - Rebased to v4.12-rc1

v4:
 - use memset to clear rwpf->mem structures

v5:
 - Rebase to v4.17-rc5
 - Rebased on top of vsp1/du/interlaced series
 - Utilise vsp1_feature() macro

v6:
 - Rebase to v4.19-rc2
 - Periods at the end of sentences.