v4l: vsp1: Use local display list for DRM flush
With the upcoming changes to the DRM VSP1 pipeline, the pipe->dl will be
committed by an asynchronous interrupt context. If this occurs during
the processing of the atomic flush then configuration of pipe->dl will be
set NULL. Rather than lock the entire function with pipe->irqlock which
brings its own difficulties, we can configure into a locally held
display list and commit atomically.

Currently the pipe->dl is set during the atomic begin hook, but it is
not utilised until the flush. Moving this should do no harm.

Signed-off-by: Kieran Bingham <kieran.bingham+renesas@ideasonboard.com>