v4l: vsp1: Provide display list body rewrite functionality
If we wish to re-use a display list or it's body we may want to adjust
parameters that have already been set in that list such as memory
addresses which would need to be updated.

Provide functionality to allow updating register values instead of
adding new ones. If the register to be updated does not exist in the
display list body, it will be added.

Using the _rewrite() API on non-reusable display lists is safe but has a
performance penalty as the existing list will be searched unnecessarily

Signed-off-by: Kieran Bingham <kieran.bingham+renesas@ideasonboard.com>