v4l: vsp1: Provide a reusable display list API
Extend the display list manager to store a separate list of display
lists which upon completion will not be cleared down.

Instead display lists on the reusable list can be obtained and
manipulated using the vsp1_dl_list_rewrite() interface.

Cooperation will be required by the consumers of the reusable display
lists to be aware of the content that must be updated.

Signed-off-by: Kieran Bingham <kieran.bingham+renesas@ideasonboard.com>