media: vsp1: Provide partition overlap algorithm
To improve image quality when scaling using the UDS we need to correctly
determine the start phase value for each partition window, and apply a
margin to overlap discontinous pixels.

Provide helper functions for calculating the phase parameters, and source
locations for a given output position and use these values to calculate our
parition window parameters.

Extend the partition algorithm to sweep first backwards, then forwards
through the entity list. Each entity is given the opportunity to expand it's window on the reverse sweep, and clip
or increase the offset on the forwards sweep.

Signed-off-by: Kieran Bingham <kieran.bingham+renesas@ideasonboard.com>

---
v2:
 - Configure HSTP and HEDP in uds_configure_partition for single partitions
 - refactored to use individual functions for various phase and position calculations
 - squashed forwards and backwards propagation work to a single patch
 - Fixed a few 'off-by-ones'
 - considerable other changes :)