media: vsp1: Implement left partition algorithm overlapping
The partition algorithm results in entities being asked to process
partition windows without full access to pixels in adjacent windows.
This can cause pixel errors on the boundaries as the filters do not have
full access to the data they require to generate the new pixel.

Allow the UDS to request an increased partition size to account for the
needs of it's internal filters, and pass this window down to the
preceding entities.

The increased window must be clipped by the WPF. This is currently
handled by setting the WPF's offset in the partition window directly
from the UDS entity, rather than propogating the offset forwards through
the entities list which would require modification to the .partition()
call.

Signed-off-by: Kieran Bingham <kieran.bingham+renesas@ideasonboard.com>

---
Setting the WPF offset directly is a bit of a hacky shortcut, as it's
the only entity which can crop pixels from the left edge. In reality I
think we should propogate the window forwards again through the pipeline
so that for example the SRU if connected after the UDS will become aware
that it has more pixels to process.