DNI/TEST: Fake long processing during atomic flush
Perform a sleep operation before processing the display list to imitate
a long delay from userspace.

This delay can be manipulated through sysfs by sending a number,
representing the quantity of milliseconds to delay as follows:

  echo 8 > /sys/module/vsp1/parameters/vsp1_delay

For testing, a value of up to 20 milliseconds is recommended, but
extreme values can be tested also.

Please note that values above 35 milliseconds, will cause faults at
pipeline shutdown where timeout values will be exceeded.

In the event of testing longer delays (such as 100 milliseconds), please
ensure that the delay is reduced before ending the stream.

Signed-off-by: Kieran Bingham <kieran.bingham+renesas@ideasonboard.com>