adv7482: Remove hack_is_hdmi() and unify DT
Now that we have multiple subdevs able to register with their own
context structures, and they are registered successfully against the
state rather than the i2c_client, we can remove the 'hack' and enable
the dual sub-node registration.

Rcar VIN will find all subdevs registered that match the base node for
the video receiver currently, so this means that ANY subdev registered
in the adv7482 will be picked up and registered by the RCarVIN driver.

This may not be correct, as we wouldn't expect input subdevs to the
Adv7482 to be registered directly by the RCarVin.

  hdmi ----\      /------ TXA ----- CSI ----- RCarVIN
            ADV7482
  cvbs ----/      \------ TXB ----- CSI ----- RcarVIN

Currently, TXA/HDMI is represented as a single subdev/entity, and
TXB/cvbs is represented as a separate single subdev/entity.

The links are only parsed for one of the entities though.

Niklas's idea of handling the links in a .registered callback seems
reasonable, except for the fact that we will get N callbacks, where N is
the number of entities/subdevs. And we only want to setup the links
once! (Not an unsolvable problem, just an untidy one)

Signed-off-by: Kieran Bingham <kieran.bingham+renesas@ideasonboard.com>