arm64: dts: r8a7795: salvator-x: Define CVBS and HDMI inputs
The Salvator-X has an HDMI and a CVBS connector on the board, which are
wired to their respective ports on the ADV7482. The CVBS is wired
directly to the AIN8 input, whilst  the HDMI is connected directly to
the input.

Provide connector representations for the CVBS and HDMI connectors in
the device tree.

Signed-off-by: Kieran Bingham <kieran.bingham+renesas@ideasonboard.com>