rcar-vin: group: use correct of_node
The unbind function dereferences the subdev->dev node to obtain the
of_node. In error paths, the subdev->dev can be set to NULL, whilst the
correct reference to the of_node is available as subdev->of_node.

Correct the dereferencing, and move the variable outside of the loop as
it is constant against the subdev, and not initialised per CSI, for both
the bind and unbind functions

Signed-off-by: Kieran Bingham <kieran.bingham+renesas@ideasonboard.com>