v4l: rdacm20: Set I2C addresses of MAX9271 and OV10635 to unique values
Change the address of the MAX9271 and OV10635 from their boot time value
to unique values provided through DT to avoid address conflicts between
multiple channels of a MAX9286 or between multiple MAX9286. I2C address
translation isn't needed with that setup, so we don't configure it.

We have to remove the writes to the OV10635 address register from its
initialization array, otherwise it will be reset to the default address.

Signed-off-by: Laurent Pinchart <laurent.pinchart+renesas@ideasonboard.com>