v4l: max9286: Map the I2C mux channel to GMSL link
I2C mux channels are numbered continuously from 0, while there could be
holes in the list of sources. Compute the mux channel number to link
number mapping at probe time and store it in an array to be used when
configuring the multiplexer.

While at it, also compute the mask of enabled sources, and use it to
setup the hardware link mask instead of erroneously computing a mask
from the number of available sources.

Signed-off-by: Laurent Pinchart <laurent.pinchart+renesas@ideasonboard.com>