v4l: max9286: Enable regulators after disabling I2C muxes
Ensure that the devices are powered in a clean compartmentalised state
by only powering up the bus in the init stage after all max9286 devices
have disabled their I2C mux.

This ensures that the cameras boot correctly in a defined state and
allows the RDACM20 driver to configure and probe successfully.

At this point we can successfully enumerate all 8 cameras and establish
a media graph consisting of both MAX9286 devices fully populated with
their subdevices.

Presently video can be captured only from channel 0 of each MAX9286, but
we can do so on both CSI busses when only the first camera is enabled on
each MAX9286.

Suggested-by: Laurent Pinchart <laurent.pinchart+renesas@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham+renesas@ideasonboard.com>