v4l: rdacm20: read device id after configuration link is enabled
I'm not sure why but some cameras always fail to probe if the driver
tries to read the device ID before first writing to the 0x04 register. I
have no idea why this is, maybe the MCU dose something funny but that
seems unlikely. This should most likely be reverted once the true
problem is found, but for now keep it.

Signed-off-by: Niklas Söderlund <niklas.soderlund+renesas@ragnatech.se>